+++ contracts/mocks/token/ERC20/ERC20FullMock.sol
@@ -5,22 +5,13 @@ import "../../../token/ERC20/IERC20Detailed.sol";
 
 contract ERC20FullMock is ERC20Full {
 
+    string public override name = "ERC20Full";
+    string public override symbol = "E2F";
+    uint8 public override decimals = 18;
+
     constructor (uint256 initialBalance) public ERC20Full(initialBalance) {}
 
     function underscoreApprove(address owner, address spender, uint256 value) public {
         super._approve(owner, spender, value);
     }
-
-    function name() external override view returns (string memory) {
-        return "ERC20Full";
-    }
-
-    function symbol() external override view returns (string memory) {
-        return "E2F";
-    }
-
-    function decimals() external override view returns (uint8) {
-        return 18;
-    }
-
 }