+++ contracts/token/ERC20/ERC20.sol
@@ -29,9 +29,9 @@ pragma solidity ^0.6.6;
 
 import "@openzeppelin/contracts/GSN/Context.sol";
 import "@openzeppelin/contracts/introspection/IERC165.sol";
-import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "@openzeppelin/contracts/math/SafeMath.sol";
 import "@openzeppelin/contracts/utils/Address.sol";
+import "./IERC20.sol";
 
 /**
  * @dev Implementation of the {IERC20} interface.